+++ src/horizontal.rs
@@ -37,7 +37,7 @@ impl HorizontalLayout {
     pub fn new() -> Self {
         Self::default()
     }
-    
+
     pub fn set_id(mut self, id: GlobalId) -> Self {
         self.id = id;
         self