+++ src/horizontal.rs
@@ -38,6 +38,11 @@ impl HorizontalLayout {
         Self::default()
     }
 
+    pub fn set_id(mut self, id: GlobalId) -> Self {
+        self.id = id;
+        self
+    }
+
     /// Appends a [`Layout`] node to the list of children.
     ///
     /// # Example