+++ CHANGELOG.md
@@ -5,4 +5,5 @@ All notable changes to agape will be documented in this file.
 ## 0.2.0 (unreleased)
 
 - Make all layout fields private ([#8](https://github.com/snubwoody/cascada/pull/8)).
-- `BlockLayout::new` now takes in a `Layout` instead of a `Box<dyn Layout>`;
+- Add `GlobalId` ([#18](https://github.com/snubwoody/cascada/pull/18)).
+- `BlockLayout::new` now takes in a `Layout` instead of a `Box<dyn Layout>` ([#17](https://github.com/snubwoody/cascada/pull/17));