+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 All notable changes to agape will be documented in this file.
 
-## 0.2.0 (unreleased)
+## 0.2.0 - 27-10-2025
 
 - Make all layout fields private ([#8](https://github.com/snubwoody/cascada/pull/8)).
 - Add `GlobalId` ([#18](https://github.com/snubwoody/cascada/pull/18)).